requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.55.3
tiktoken==0.5.2

# Database
faiss-gpu==1.7.2
# For CPU-only systems, use: faiss-cpu==1.7.4

# Data processing
pandas==2.1.4
numpy==1.26.2

# NLP
nltk==3.8.1

# ML/Similarity
scikit-learn==1.3.2
scipy==1.11.4

# Environment
python-dotenv==1.0.0

# SQLAlchemy for database
sqlalchemy==2.0.23